refactor: 整理 package
Closes #16
