+++ backend/src/main/java/tw/waterballsa/gaas/unoflip/UnoFlipApplication.java
@@ -4,15 +4,13 @@ import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-@EnableScheduling
 @OpenAPIDefinition(info = @Info(title = "Uno Flip API", version = "1.0.0"))
 public class UnoFlipApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(UnoFlipApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(UnoFlipApplication.class, args);
+	}
 
 }
