+++ backend/src/main/java/tw/waterballsa/gaas/unoflip/response/JoinResult.java
@@ -1,4 +1,4 @@
-package tw.waterballsa.gaas.unoflip.vo;
+package tw.waterballsa.gaas.unoflip.response;
 
 import tw.waterballsa.gaas.unoflip.domain.PlayerInfo;
