+++ backend/src/main/java/tw/waterballsa/gaas/unoflip/response/Response.java
@@ -1,4 +1,4 @@
-package tw.waterballsa.gaas.unoflip.vo;
+package tw.waterballsa.gaas.unoflip.response;
 
 public record Response<T>(int code, String message, T payload) {
 }
