+++ backend/src/main/java/tw/waterballsa/gaas/unoflip/vo/GameJoinResult.java
@@ -4,5 +4,5 @@ import tw.waterballsa.gaas.unoflip.domain.PlayerInfo;
 
 import java.util.List;
 
-public record GameJoinResult (int tableId, List<PlayerInfo> playerInfoList){
+public record GameJoinResult(int tableId, List<PlayerInfo> playerInfoList) {
 }
