+++ backend/src/test/java/tw/waterballsa/gaas/unoflip/presenter/GameJoinPresenterTest.java
@@ -5,7 +5,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tw.waterballsa.gaas.unoflip.domain.PlayerInfo;
 import tw.waterballsa.gaas.unoflip.event.BroadcastEvent;
+import tw.waterballsa.gaas.unoflip.event.EventType;
 import tw.waterballsa.gaas.unoflip.event.JoinBroadcastEvent;
+import tw.waterballsa.gaas.unoflip.response.JoinResult;
+import tw.waterballsa.gaas.unoflip.response.Response;
 import tw.waterballsa.gaas.unoflip.vo.*;
 
 import java.util.Arrays;
